\name{CoxKL-package}
\alias{CoxKL-package}
\alias{CoxKL}
\docType{package}
\title{
  Kullback-Leibler-Based Cox Models
}
\description{
  Incorporating external information can improve the performance of risk discrimination in an internal small-sized cohort. However, given that external risk scores and internal individual-level data come from different populations, ignoring population heterogeneity can introduce substantial bias. Thus, we develop a Kullback-Leibler-based Cox model (CoxKL) to integrate internal individual-level time-to-event data with external risk scores derived from published prediction models. 
}
\author{
Di Wang

Maintainer: Di Wang <dwwang@umich.edu>
}

